still need to add delete and fix url parameter for delete
