+++ questions/Pages/Shared/_AlertMessage.cshtml
@@ -1,0 +1,12 @@
+﻿@if (TempData["message"] != null)
+{
+    <div class="col-md-12">
+        <div class="@(TempData["class"] + " " + "alert-dismissible fade show")" role="alert">
+            @TempData["message"]
+            <button type="button" class="close" data-dismiss="alert" data-bs-dismiss="alert" aria-label="Close">
+                <span aria-hidden="true">&times;</span>
+            </button>
+        </div>
+    </div>
+    
+}
