+++ questions/Views/ExamRepo/Index.cshtml
@@ -19,15 +19,7 @@
         @ViewData["Title"]
         <hr />
     </div>
-    @if(TempData["message"] != null)
-    {
-        <div class="@(TempData["class"] + "alert-dismissible fade show")" role="alert">
-            @TempData["message"]
-            <button type="button" class="close" data-dismiss="alert" aria-label="Close">
-                <span aria-hidden="true">&times;</span>
-            </button>
-        </div>
-    }
+    @await Html.PartialAsync("_AlertMessage")
     <div class="col-md-12">
         <a href="@Url.Action("Repository")" class="btn btn-success">Add <i class="bi bi-plus"></i></a>
     </div>
