+++ questions/Views/ExamRepo/Repository.cshtml
@@ -1,23 +1,47 @@
-﻿@*
+﻿@using questions.Models
+@*
     For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 *@
+@model ExamRepoPostVM
 @{
     Layout = "_Layout";
-    ViewData["Title"] = "Exam Repositories - Edit";
+    if (ViewBag.RepoId == null)
+    {
+        ViewData["Title"] = "Exam Repositories - Add";
+    }
+    else
+    {
+        ViewData["Title"] = "Exam Repositories - Edit";
+    }
 
 }
 
-
 <div class="row">
     <div class="col-md-12">
-        test
+        @ViewData["Title"]
+        <hr />
     </div>
+    @await Html.PartialAsync("_AlertMessage")
+</div>
+<div class="row">
+    
+        @using (Html.BeginForm("Repository", "ExamRepo",FormMethod.Post))
+        {
+            @Html.HiddenFor(m=>m.RepoId)
+            <div class="col-md-12 form-group">
+                @Html.LabelFor(f=> f.Name,new {@class="control-label"})
+                @Html.TextAreaFor(f=>f.Name,5,5,new {@class="form-control"})
+                @Html.ValidationMessageFor(i => i.Name, "", new { @class = "text-danger" })
+
+            </div>
+            <div class="col-md-12 form-group">
+            @Html.ValidationSummary(false,"", new { @class = "text-danger" })
+            </div>
+            <div class="col-md-12 form-group">
+                <button type="submit" class="btn btn-success">Save</button>
+            </div>
+        }
+    
 </div>
 
 
-
-<div class="row">
-    <div class="col-md-12">
-        <a class="btn btn-warning" href="@Url.Action("Index")">Back</a>
-    </div>
-</div>
